Remove check that blocked group details for new users
GroupInfo.vue doesn't let a user see any group's detail page if the user
doesn't belong to at least one group. However, users without a group can
still navigate to the group detail page (via direct link or the group
overview page, i.e. `/dashboard/groups`) and will then only see a blank
page.

This commit fixes this behavior by entirely removing the GroupInfo.vue
component. Its only functions were to perform the check mentioned above
and to set the search mode to "group". It is not used anywhere else, the
GroupListing component is a drop-in replacement for the route belonging
to GroupInfo and the search mode can be set by the former as well.

[Issue: #228]