Fix TypeError if the current group isn't found
The `currentGroup` variable in the vuex state is set to a group object
from one of the `userGroups` or `publicGroups` variables. Usually these
are already filled with the corresponding groups because the user has
been on a page that has loaded these variables, but if the group detail
page is reloaded, these variables aren't filled, `currentGroup` is set
to null and later accessed which then causes a TypeError.

This commit fixes this possible error by ensuring that the `userGroups`
and `publicGroups` variables have been filled before setting the
`currentGroup` variable.

[Issue: #229]