Conditionally disable file dropping on group pages
If the logged-in user isn't allowed to add panels to the group they are
viewing they are now unable to drag-and-drop files on the page. This was
not a security issue, as the file that is dropped is uploaded but not
added to the current group. UX-wise, though, this change should remove
any confusion over what happens with the uploaded panel.

[Issue: #202]