run check for logged-in user before mounting Vue
The Vue router route-guards were running earlier in the application
lifecycle than the application created event, where the check for the
logged in user was performed. This commit moves the check for a logged
in user to occur before the Vue application is mounted, so the
application is aware of whether there is a logged-in user.