fix errors on panel duplication tests
The errors were caused by the User seeder randomly allocating roles to
the generated user. Because a user was occasionally allocated the role
of superuser, they were able to carry out tasks that their role at the
panel level should not allow. Therefore, all users generated in the
duplication tests are specified as regular users.